committing bc going to bed

+++ recipe/models.py
@@ -4,13 +4,13 @@ from django.db import models
 class Menu(models.Model):
     name = models.CharField(max_length=50)
     description = models.CharField(max_length=255)
-    items = models.ManyToManyField('MenuItem', related_name='menus')
+    items = models.ManyToManyField("MenuItem", related_name="menus")
 
 
 class MenuItem(models.Model):
     name = models.CharField(max_length=50)
     description = models.CharField(max_length=255)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
 
 
 class Recipe(models.Model):
@@ -23,7 +23,5 @@ class Ingredient(models.Model):
     quantity = models.IntegerField()
     unit = models.CharField(max_length=100)
     recipe = models.ForeignKey(
-        Recipe,
-        on_delete=models.CASCADE,
-        related_name='ingredients'
+        Recipe, on_delete=models.CASCADE, related_name="ingredients"
     )
